Move directory to users package
