+++ common/test/python/test_directory.py
@@ -2,8 +2,8 @@
 from datetime import datetime
 
 import yaml
-from redcap.nacc_directory import (Authorizations, Credentials, PersonName,
-                                   UserDirectoryEntry)
+from users.nacc_directory import (Authorizations, Credentials, PersonName,
+                                  UserDirectoryEntry)
 
 
 # pylint: disable=(no-self-use,too-few-public-methods)
